Add :element option
