+++ lib/capybara/node/elements/field.rb
@@ -1,0 +1,9 @@
+module Capybara
+  module Node
+    module Elements
+      class Field < Element
+        class_for :field
+      end
+    end
+  end
+end
