+++ lib/capybara/node/elements/select.rb
@@ -1,0 +1,9 @@
+module Capybara
+  module Node
+    module Elements
+      class Select < Element
+        class_for :select
+      end
+    end
+  end
+end
